Feat subir evidencia pago

+++ src/main/java/pe/edu/utp/dwi/HBSGool/reservacion/ReservacionDto.java
@@ -14,7 +14,7 @@ public class ReservacionDto {
 
     private Integer idReservacion;
     private Integer usuarioId;
-    private Short canchaId;
+    private Integer canchaId;
     private Short cajeroId;
     private LocalDateTime tiempoInicio;
     private String dni;
