+++ src/main/java/pe/edu/utp/dwi/HBSGool/reservacion/ReservacionEntity.java
@@ -26,7 +26,7 @@ public class ReservacionEntity {
 
     private Integer usuarioId;
 
-    private Short canchaId;
+    private Integer canchaId;
 
     private Short cajeroId;
